Fix Oracle issue in gene name constraint.
Oracle limits the number of elements in an 'in' constraint.
The gene name constraint therefore needed to be adapted to
split large query constraints into smaller constraints.
